+++ R_tasks/prepare_BNT.R
@@ -1,4 +1,5 @@
-##' Prepare BNT
+##' Prepare BNT 
+##' Should work with BNT and BNTen
 ##'
 ##' Template for the functions to prepare specific tasks. Most of this file should not be changed
 ##' Things to change: 
@@ -17,18 +18,19 @@
 prepare_BNT <- function(DF_clean, short_name_scale_str) {
 
   # DEBUG
+  # targets::tar_load_globals()
   # debug_function(prepare_BNT)
 
   # Standardized names ------------------------------------------------------
   names_list = standardized_names(short_name_scale = short_name_scale_str, 
-                     # dimensions = c("NameDimension1", "NameDimension2"), # Use names of dimensions, "" or comment out line
                      help_names = FALSE) # help_names = FALSE once the script is ready
   
   # Create long -------------------------------------------------------------
   DF_long_RAW = create_raw_long(DF_clean, short_name_scale = short_name_scale_str, numeric_responses = TRUE, help_prepare = FALSE) %>% 
     
     # In BNT we have exit conditions. Add "BNT_03","BNT_04" to make we have all the columns we need to complete the correction
-    dplyr::bind_rows(tibble(trialid = c("BNT_03","BNT_04"))) %>% 
+      # short_name_scale_str can be BNT or BNTen
+    dplyr::bind_rows(tibble(trialid = c(paste0(short_name_scale_str, "_03"),paste0(short_name_scale_str, "_04")))) %>% 
     tidyr::complete(trialid, nesting(id, experiment)) %>% 
     tidyr::drop_na(id) 
   
@@ -57,12 +59,12 @@ prepare_BNT <- function(DF_clean, short_name_scale_str) {
     dplyr::mutate(
       DIR =
        dplyr::case_when(
-          # trialid == "BNT_01" & as.numeric(RAW) >= 400 & as.numeric(RAW) <= 600 ~ 1,
-          trialid == "BNT_02" & RAW != 30 ~ 1,
-          trialid == "BNT_02" & RAW == 30 ~ 2,
-          trialid == "BNT_03" & RAW == 20 ~ 4,
-          trialid == "BNT_04" & RAW == 50 ~ 4,
-          trialid == "BNT_04" & RAW != 50 ~ 3,
+          # trialid == "BNTen_01" & as.numeric(RAW) >= 400 & as.numeric(RAW) <= 600 ~ 1,
+          trialid %in% c("BNT_02", "BNTen_02") & RAW != 30 ~ 1,
+          trialid %in% c("BNT_02", "BNTen_02") & RAW == 30 ~ 2,
+          trialid %in% c("BNT_03", "BNTen_03") & RAW == 20 ~ 4,
+          trialid %in% c("BNT_04", "BNTen_04") & RAW == 50 ~ 4,
+          trialid %in% c("BNT_04", "BNTen_04") & RAW != 50 ~ 3,
           is.na(RAW) ~ NA_real_,
           grepl(items_to_ignore, trialid) ~ NA_real_,
           TRUE ~ 0
@@ -82,14 +84,12 @@ prepare_BNT <- function(DF_clean, short_name_scale_str) {
       values_from = c(RAW, DIR),
       names_glue = "{trialid}_{.value}") %>% 
     
-    # ENMIENDA MANUAL para corregir error en piloto. Si han acertado el BNT_03_DIR, no deberian ver el BNT_04_DIR ------------------
-    dplyr::mutate(BNT_04_DIR = 
-            dplyr::case_when(
-               BNT_03_DIR == 4 & BNT_04_DIR == 4 ~ 0,
-               TRUE ~ BNT_04_DIR
-             )) %>% 
-    
-    
+    # # ENMIENDA MANUAL para corregir error en piloto. Si han acertado el BNT_03_DIR, no deberian ver el BNT_04_DIR ------------------
+    # dplyr::mutate(BNT_04_DIR = 
+    #         dplyr::case_when(
+    #            BNT_03_DIR == 4 & BNT_04_DIR == 4 ~ 0,
+    #            TRUE ~ BNT_04_DIR
+    #          )) %>% 
     
     
     # NAs for RAW and DIR items
@@ -102,14 +102,8 @@ prepare_BNT <- function(DF_clean, short_name_scale_str) {
     # [USE STANDARD NAMES FOR Scales and dimensions: name_DIRt, name_DIRd1, etc.] Check with: standardized_names(help_names = TRUE)
 
     dplyr::mutate(
-
-      # Score Dimensions (see standardized_names(help_names = TRUE) for instructions)
-      # !!names_list$name_DIRd[1] := rowSums(select(., matches("02|04|05") & matches("_DIR$")), na.rm = TRUE), 
-      # !!names_list$name_DIRd[2] := rowSums(select(., matches("01|03|08") & matches("_DIR$")), na.rm = TRUE), 
-      
       # Score Scale
       !!names_list$name_DIRt := rowSums(select(., matches("_DIR$")), na.rm = TRUE)
-      
     )