+++ _targets_options.R
@@ -26,7 +26,7 @@ essential_packages = c(
 
 extra_packages = c(
                   # Extras
-                  "furrr", "patchwork", "writexl", "broom", "broom.mixed", "emmeans", "gt", "irr", "lme4", "parameters", "performance", "ggalluvial", "ggridges",
+                  "furrr", "writexl", "broom", "broom.mixed", "emmeans", "gt", "irr", "lme4", "parameters", "performance", "ggalluvial", "ggridges",
                   
                   # Non declared dependencies
                   "diffviewer", "qs", "webshot", "xml2", "jquerylib",